# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "CinemaVault API",
    "version": "1.0.0",
    "description": "API documentation for the CinemaVault film discovery platform."
  },
  "host": "localhost:5000",
  "basePath": "/api",
  "schemes": ["http"],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register a new platform administrator",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": { "type": "string" },
                "password": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "201": { "description": "Admin registered successfully" },
          "400": { "description": "Invalid data or user exists" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Admin login to receive JWT Token",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": { "type": "string" },
                "password": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "Login successful, returns token" },
          "400": { "description": "Invalid credentials" }
        }
      }
    },
    "/movies": {
      "get": {
        "summary": "Public browse, search, and filter movies",
        "parameters": [
          { "name": "title", "in": "query", "type": "string", "required": false },
          { "name": "genre", "in": "query", "type": "string", "required": false },
          { "name": "year", "in": "query", "type": "string", "required": false }
        ],
        "responses": {
          "200": { "description": "A list of movies" }
        }
      },
      "post": {
        "summary": "Add a new movie (Admin Only)",
        "parameters": [
          { "name": "Authorization", "in": "header", "type": "string", "required": true, "description": "Bearer <JWT_TOKEN>" },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "title": { "type": "string" },
                "genre": { "type": "string" },
                "year": { "type": "string" },
                "rating": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "201": { "description": "Movie created" },
          "401": { "description": "Unauthorized" }
        }
      }
    }
  }
}